pyproject.toml
[project]
name = "BigBunther"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Private Butts", email = "randybuttsii@gmail.com"},
]
dependencies = [
    "structlog>=24.1.0",
    "django-environ>=0.11.2",
    "aiohttp[speedups]>=3.9.5",
    "pillow>=10.3.0",
    "discord-py>=2.3.2",
    "python-ffmpeg>=2.0.12",
    "pygifsicle>=1.0.7",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "watchdog>=4.0.0",
    "debugpy>=1.8.1",
]
